Separated common code from read kernel.
